[server] Fix compile errors from performance package migration
Client was missing the Body field used by mesheryctl when parsing a
performance profile, and testClient.Rps is now int32 so it needed an
explicit cast before FormatInt. Also ran gofmt on the files touched by
the service-mesh-performance removal to fix import grouping and struct
tag alignment.

+++ mesheryctl/internal/cli/root/perf/apply.go
@@ -144,7 +144,7 @@ mesheryctl perf apply meshery-profile-new --url "https://google.com" --load-gene
 			}
 
 			if qps == "" {
-				qps = strconv.FormatInt(testClient.Rps, 10)
+				qps = strconv.FormatInt(int64(testClient.Rps), 10)
 			}
 
 			if concurrentRequests == "" {
